refactor: separate extension runtime staging
