+++ package.json
@@ -64,24 +64,30 @@
 		"packages/plugin-godot/assets/godot.svg",
 		"packages/plugin-godot/dist/extension.js",
 		"packages/plugin-godot/dist/plugin.js",
+		"packages/plugin-godot/LICENSE",
 		"packages/plugin-godot/package.json",
 		"packages/plugin-markdown/dist/plugin.js",
+		"packages/plugin-markdown/LICENSE",
 		"packages/plugin-markdown/package.json",
 		"packages/plugin-particles/dist/plugin.js",
 		"packages/plugin-particles/dist/webview.js",
+		"packages/plugin-particles/LICENSE",
 		"packages/plugin-particles/package.json",
 		"packages/plugin-svelte/dist/plugin.js",
+		"packages/plugin-svelte/LICENSE",
 		"packages/plugin-svelte/package.json",
 		"packages/plugin-typescript/dist/plugin.js",
+		"packages/plugin-typescript/LICENSE",
 		"packages/plugin-typescript/package.json",
 		"packages/plugin-unity/assets/NOTICE.md",
 		"packages/plugin-unity/assets/unity.svg",
 		"packages/plugin-unity/dist/extension.js",
 		"packages/plugin-unity/dist/plugin.js",
+		"packages/plugin-unity/LICENSE",
 		"packages/plugin-unity/package.json",
 		"packages/plugin-vue/dist/plugin.js",
+		"packages/plugin-vue/LICENSE",
 		"packages/plugin-vue/package.json",
-		"docs/**",
 		"README.md",
 		"CONTRIBUTING.md",
 		"LICENSE"
